+++ lib/builder.js
@@ -1,155 +1,313 @@
-var saucy = require('./sourcemaps');
-var path = require('path');
-var mkdirp = require('mkdirp');
+var Promise = require('rsvp').Promise;
+var System = require('systemjs');
+
+var asp = require('rsvp').denodeify;
 var fs = require('fs');
-var rsvp = require('rsvp');
-var asp = rsvp.denodeify;
-var uglify = require('uglify-js');
+var path = require('path');
 
-function countLines(str) {
-  return str.split(/\r\n|\r|\n/).length;
-}
+var attachCompilers = require('./compile').attachCompilers;
+var compileOutputs = require('./compile').compileOutputs;
+var writeOutputs = require('./output').writeOutputs;
 
-// Process compiler outputs, gathering:
-//
-//   concatOutputs:         list of source strings to concatenate
-//   sourceMapsWithOffsets: list of [absolute offset,
-//                                   source map string] pairs
-//
-//  Takes lists as empty references and populates via push.
-function processOutputs(outputs) {
-  var offset = 0;
 
-  var outputObj = {};
-
-  var sources = outputObj.sources = [];
-  var sourceMapsWithOffsets = outputObj.sourceMapsWithOffsets = [];
-
-  if (!outputs)
-    return outputObj;
-
-  outputs.forEach(function(output) {
-    var source;
-    if (typeof output == 'object') {
-      source = output.source || '';
-      var offset_ = output.sourceMapOffset || 0;
-      var map = output.sourceMap;
-      if (map) {
-        sourceMapsWithOffsets.push([offset + offset_, map]);
-      }
-    }
-    else if (typeof output == 'string') {
-      source = output;
-    }
-    else {
-      throw "Unexpected output format: " + output.toString();
-    }
-    source = saucy.removeSourceMaps(source || '');
-    offset += countLines(source);
-    sources.push(source);
-  });
-
-  return outputObj;
-}
-
-function createOutput(outputs, outFile, baseURL, opts) {
-  // process output
-  var sourceMap;
-
-  var outputObj = processOutputs(outputs);
-
-  if (opts.sourceMaps && outputObj.sourceMapsWithOffsets.length) {
-    var sourceRoot = outFile ? path.dirname(baseURL + outFile) + path.sep : baseURL;
-    var mapsWithOffsets = outputObj.sourceMapsWithOffsets;
-    sourceMap = saucy.concatenateSourceMaps(outFile, mapsWithOffsets, sourceRoot, opts.sourceMapContents);
-  }
-
-  var output = outputObj.sources.join('\n');
-
-  return {
-    source: output,
-    sourceMap: sourceMap
+function processOpts(opts_, outFile) {
+  var opts = {
+    config: {},
+    lowResSourceMaps: false,
+    minify: false,
+    normalize: false,
+    outFile: outFile,
+    runtime: false,
+    sourceMaps: false,
+    sourceMapContents: opts_ && opts_.sourceMaps == 'inline'
   };
+  for (var key in opts_) {
+    if (key in opts)
+      opts[key] = opts_[key];
+  }
+  return opts;
 }
 
-function minify(output, fileName, mangle) {
-  var ast = uglify.parse(output.source, { filename: fileName });
-  ast.figure_out_scope();
-  ast = ast.transform(uglify.Compressor({ warnings: false }));
-  ast.figure_out_scope();
-  ast.compute_char_frequency();
-  if (mangle !== false) {
-    ast.mangle_names({
-      except: ['require']
-    });
+function Builder(cfg) {
+  this.loader = null;
+  this.reset();
+  if (typeof cfg == 'string')
+    this.loadConfigSync(cfg);
+  else if (typeof cfg == 'object')
+    this.config(cfg);
+}
+
+Builder.prototype.reset = function() {
+  var loader = this.loader = System.clone();
+  loader.baseURL = System.baseURL;
+
+  var pluginLoader = System.clone();
+  pluginLoader.baseURL = System.baseURL;
+  pluginLoader.trace = true;
+
+  loader.trace = true;
+  loader.execute = false;
+  loader.pluginLoader = pluginLoader;
+
+  attachCompilers(loader);
+};
+
+function executeConfigFile(loader, source) {
+  var curSystem = global.System;
+  global.System = {
+    config: function(cfg) {
+      loader.config(cfg);
+      loader.pluginLoader.config(cfg);
+    }
+  };
+  // jshint evil:true
+  new Function(source.toString()).call(global);
+  global.System = curSystem;
+}
+
+var resolvePath = path.resolve.bind(path, process.cwd());
+
+Builder.prototype.loadConfig = function(configFile) {
+  var self = this;
+  return asp(fs.readFile)(resolvePath(configFile))
+    .then(executeConfigFile.bind(null, this.loader))
+    .then(function() { return self; });
+};
+
+Builder.prototype.loadConfigSync = function(configFile) {
+  var source = fs.readFileSync(resolvePath(configFile));
+  executeConfigFile(this.loader, source);
+};
+
+Builder.prototype.config = function(config) {
+  var loader = this.loader;
+  var pluginLoader = loader.pluginLoader;
+
+  var cfg = {};
+  for (var p in config) {
+    if (p != 'bundles')
+      cfg[p] = config[p];
   }
+  loader.config(cfg);
+  pluginLoader.config(cfg);
+};
 
-  var sourceMap;
-  if (output.sourceMap)
-    sourceMap = uglify.SourceMap({
-      file: fileName,
-      orig: output.sourceMap
-    });
+Builder.prototype.build = function(moduleName, outFile, opts) {
+  var self = this;
+  opts = opts || {};
 
-  output.source = ast.print_to_string({
-    // for some reason non-ascii broke esprima.... this does break unicode though
-    ascii_only: true,
-    // keep first comment
-    comments: function(node, comment) {
-      return comment.line === 1 && comment.col === 0;
-    },
-    source_map: sourceMap
+  return this.trace(moduleName, opts.config)
+  .then(function(trace) {
+    return self.buildTree(trace.tree, outFile, opts);
   });
-  output.sourceMap = sourceMap;
+};
 
-  return output;
-}
+function parseExpression(expressionString) {
+  var args = expressionString.split(' ');
 
-function writeOutputFile(opts, output, basePath) {
-  var sourceMapFile;
-  if (opts.sourceMaps && output.sourceMap) {
-    sourceMapFile = path.basename(opts.outFile) + '.map';
-    output.source += '\n//# sourceMappingURL=' + sourceMapFile;
+  var firstModule = args[0];
+
+  var operations = [];
+
+  for (var i = 1; i < args.length - 1; i = i + 2) {
+    var operator = args[i];
+    var moduleName = args[i + 1];
+
+    operations.push({
+      operator: operator,
+      moduleName: moduleName
+    });
   }
 
-  return asp(mkdirp)(path.dirname(path.resolve(basePath, opts.outFile)))
+  return {firstModule: firstModule, operations: operations};
+};
+
+function lookupOperatorFn(symbol) {
+  if (symbol == '+')
+    return this.addTrees;
+  else if (symbol == '-')
+    return this.subtractTrees;
+  else
+    throw 'Unknown operator ' + op.operator;
+};
+
+Builder.prototype.buildExpression = function(expression, cfg) {
+  var builder = this;
+
+  if (typeof expression == 'string')
+    expression = parseExpression(expression);
+
+  var firstModule = expression.firstModule;
+  var operations = expression.operations;
+
+  return Promise.resolve(builder.trace(firstModule, cfg))
+  .then(function(trace) {
+    // if there are no other operations, then we have the final tree
+    if (!operations.length)
+      return trace.tree;
+
+    var applyOperation = function(promise, op) {
+      return promise.then(function(curTree) {
+        return builder.trace(op.moduleName)
+        .then(function(nextTrace) {
+          var operatorFn = lookupOperatorFn(op.operator);
+          return operatorFn.call(builder, curTree, nextTrace.tree);
+        });
+      });
+    };
+
+    // chain the operations, applying them with the trace of the next module
+    return operations.reduce(applyOperation, Promise.resolve(trace.tree));
+  });
+};
+
+Builder.prototype.buildTree = function(tree, outFile, opts) {
+  var loader = this.loader;
+  opts = processOpts(opts, outFile);
+
+  return compileOutputs(loader, tree, opts, false)
+  .then(function(outputs) {
+    return writeOutputs(opts, outputs, loader.baseURL);
+  });
+};
+
+Builder.prototype.buildSFX = function(moduleName, outFile, opts) {
+  var loader = this.loader;
+  opts = processOpts(opts, outFile);
+  opts.normalize = true;
+
+  return this.trace(moduleName, opts.config)
+  .then(function(trace) {
+    return compileOutputs(loader, trace.tree, opts, trace.moduleName);
+  })
+  .then(function(outputs) {
+    return writeOutputs(opts, outputs, loader.baseURL);
+  });
+};
+
+Builder.prototype.trace = function(moduleName, config, includePlugins) {
+  var loader = this.loader;
+  var pluginLoader = loader.pluginLoader;
+
+  if (config) {
+    this.config(config);
+  }
+
+  var System = loader.global.System;
+  loader.global.System = loader;
+
+  var traceTree = {};
+
+  return loader.import(moduleName)
   .then(function() {
-    if (!output.sourceMap || !opts.sourceMaps) return;
-    var sourceMapPath = path.resolve(basePath, path.dirname(opts.outFile), sourceMapFile);
-    return asp(fs.writeFile)(sourceMapPath, output.sourceMap);
+    return loader.normalize(moduleName);
+  })
+  .then(function(_moduleName) {
+    moduleName = _moduleName;
+    loader.global.System = System;
+    return visitTree(loader.loads, moduleName, includePlugins && pluginLoader, function(load) {
+      traceTree[load.name] = load;
+    });
   })
   .then(function() {
-    var sourcePath = path.resolve(basePath, opts.outFile);
-    return asp(fs.writeFile)(sourcePath, output.source);
+    return {
+      moduleName: moduleName,
+      tree: traceTree
+    };
+  })
+  .catch(function(e) {
+    loader.global.System = System;
+    throw e;
+  });
+};
+
+function visitTree(tree, moduleName, pluginLoader, visit, seen) {
+  seen = seen || [];
+
+  if (seen.indexOf(moduleName) != -1)
+    return;
+
+  seen.push(moduleName);
+
+  var load = tree[moduleName];
+
+  if (!load)
+    return Promise.resolve();
+
+  // visit the deps first
+  return Promise.all(load.deps.map(function(dep) {
+    return visitTree(tree, load.depMap[dep], pluginLoader, visit, seen);
+  })).then(function() {
+    if (!pluginLoader)
+      return;
+
+    var pluginName = load.metadata.pluginName;
+    if (pluginName) {
+      return visitTree(pluginLoader.loads, pluginName, pluginLoader, visit, seen);
+    }
+  })
+  .then(function() {
+    // if we are the bottom of the tree, visit
+    return visit(load);
   });
 }
 
-exports.inlineSourceMap = inlineSourceMap;
-function inlineSourceMap (output) {
-  return output.source +
-    '\n//# sourceMappingURL=data:application/json;base64,' +
-    new Buffer(output.sourceMap.toString()).toString('base64');
-}
+// returns a new tree containing tree1 n tree2
+Builder.intersectTrees = function(tree1, tree2) {
+  var name;
+  var intersectTree = {};
 
-exports.writeOutput = function(opts, outputs, baseURL) {
-  // remove 'file:' part
-  var basePath = baseURL.substr(5);
+  var tree1Names = [];
+  for (name in tree1)
+    tree1Names.push(name);
 
-  if (opts.outFile)
-    opts.outFile = path.relative(basePath, path.resolve(opts.outFile));
+  for (name in tree2) {
+    if (tree1Names.indexOf(name) == -1)
+      continue;
 
-  var output = createOutput(outputs, opts.outFile, baseURL, opts);
-
-  if (opts.minify)
-    output = minify(output, opts.outFile, opts.mangle);
-
-  if (opts.sourceMaps == 'inline') {
-    output.source = inlineSourceMap(output);
-    output.sourceMap = undefined;
+    intersectTree[name] = tree1[name];
   }
 
-  if (opts.outFile)
-    return writeOutputFile(opts, output, basePath).then(function() { return output; });
-  else
-    return rsvp.resolve(output);
+  return intersectTree;
 };
+
+// returns a new tree containing tree1 + tree2
+Builder.addTrees = function(tree1, tree2) {
+  var name;
+  var unionTree = {};
+
+  for (name in tree2)
+    unionTree[name] = tree2[name];
+
+  for (name in tree1)
+    unionTree[name] = tree1[name];
+
+  return unionTree;
+};
+
+// returns a new tree containing tree1 - tree2
+Builder.subtractTrees = function(tree1, tree2) {
+  var name;
+  var subtractTree = {};
+
+  for (name in tree1)
+    subtractTree[name] = tree1[name];
+
+  for (name in tree2)
+    delete subtractTree[name];
+
+  return subtractTree;
+};
+
+// copies a subtree out of the tree
+Builder.extractTree = function(tree, moduleName) {
+  var outTree = {};
+  return visitTree(tree, moduleName, null, function(load) {
+    outTree[load.name] = load;
+  })
+  .then(function() {
+    return outTree;
+  });
+};
+
+module.exports = Builder;
