+++ lib/compilers/cjs.js
@@ -75,7 +75,7 @@ CJSRegisterTransformer.prototype.transformScript = function(tree) {
 }
 
 exports.compile = function(load, opts, loader) {
-  var options = { script: true };
+  var options = { script: true, sourceRoot: true };
   if (opts.sourceMaps)
     options.sourceMaps = 'memory';
   if (opts.lowResSourceMaps)
