+++ lib/compilers/es6.js
@@ -68,6 +68,7 @@ exports.compile = function(load, opts, loader) {
     options = loader.traceurOptions || {};
     options.modules = 'instantiate';
     options.script = false;
+    options.sourceRoot = true;
     options.moduleName = load.name;
 
     if (opts.sourceMaps)
