+++ lib/compilers/global.js
@@ -78,7 +78,7 @@ GlobalTransformer.prototype.transformScript = function(tree) {
 }
 
 exports.compile = function(load, opts, loader) {
-  var options = { script: true };
+  var options = { script: true, sourceRoot: true };
 
   if (opts.sourceMaps)
     options.sourceMaps = 'memory';
