+++ lib/sourcemaps.js
@@ -38,7 +38,7 @@ exports.concatenateSourceMaps = function(sourceFilename, mapsWithOffsets, source
     var map = getMapObject(pair[1]);
 
     wrapSourceMap(map).eachMapping(function(mapping) {
-      if (mapping.source.match(/\/@traceur/))
+      if (mapping.source.match(/(\/|^)@traceur/))
         return;
 
       generated.addMapping({
