+++ lib/templates/sfx-core.js
