+++ test/cfg-manual.js
@@ -1,7 +1,8 @@
 System.config({
   baseURL: '',
   paths: {
-    'jquery-cdn': 'https://code.jquery.com/jquery-2.1.1.min.js'
+    'jquery-cdn': 'https://code.jquery.com/jquery-2.1.1.min.js',
+    '@babel': '../node_modules/babel-core/browser.js'
   },
   meta: {
     'jquery-cdn': {
